log out clean up ish
