+++ package.json
@@ -31,6 +31,7 @@
     "next-transpile-modules": "^7.2.0",
     "pluralize": "^8.0.0",
     "promise-queue": "^2.2.5",
+    "query-string": "^7.1.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-grid-system": "^7.1.2",
